Major clean up of the model_mod.f90 code.  Added functionality to return KIND_TEMPERATURE more efficiently from the interpolate routine if the vertical coordinate is pressure (sensible temperature is computed as a side-effect of converting pressure into height). Also added an efficient way to compute KIND_TEMPERATURE for other vertical coordinates.  Added support for vert_is_surface(). Caught vert_is_undef() and return error for now.  Reformatted routines to be consistent in spacing, where separator lines are, where comments are, where args are, etc.   Added option in namelist to specify the grid netcdf file separately from the file with the current analysis data, in the expectation that we won't want to replicate all the grid info in each ensemble member because of space considerations.
Added option to model_mod_check to return multiple vertical
slices in addition to the horizontal.

Updated namelist to match changes to model_mod and model_mod_check.

I am hoping next week to reorder the subroutines into
more logical sections and put the most used routines near
the top.  I'm not doing that now because it makes it
almost impossible to diff the changes and I want to talk
to everyone to be sure any other changes are merged in
first before moving code blocks around.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/mpas@5299 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 3dbe7245aec556fc734378540b9317bc930a9ee8