This group does not depend on the CAM pointer file, it uses logic more similar to the hybrid case to determine the cam initial filename as well as the cam history file (source of the PHIS field). It also stages a copy of the necessary DART bits in the caseroot directory - facilitating concurrent cases/experiments. The assimilate script grabs the bits from the caseroot directory instead of the DART directory.
Mariana also suggested a SourceMod that prevents the coupler from
writing a restart file. The pointer file to the coupler restart is
also not written - but was used by the short-term archive script to
create a unique directory for staging restarts. I have modified the
short-term archive script to get that information from the 
Posterior_Diag.YYYY-MM-DD-SSSSS.nc file.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/development@5508 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: e5cfe995f463ec1fc58b9b5a74fa5234c176d750