Added code to compute barycentric weights, seems to be working for scalar values but needs extensive testing.  I've broken the RBF interpolation for winds, however, apparently, since all i'm ever getting back is NaN.  will look at that next.
when matt originally gave me the rbf files, i told him the
wrong index order for the normal vector array and he 'fixed' it.
i tried to revert to the original, correct fortran array order
for the normal vectors.  those changes should be carefully checked
because they may be part of the problem.


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/mpas@5526 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: b2ef359c3412a4fed856a017e0716365e68047c3