removed the verbose flag because it doesn't control whether the message lines print or not.   in case of error added: call set_output(.true.) so all tasks will print messages, not just 0.  in an error situation it's possible for tasks other than 0 to have a different time for the data and before it would silently fail without printing the messages about why.
also rearranged the section so the logic should be easier to follow.
tested with trace_execution true and false (which controls the setting
of print_trace_details) to ensure the info gets printed in either case.
(in this code we could use an error_handler() flag which is an error
but still returns instead of exiting.  we could use it to force
printing in all cases and have multiple calls with info before the
final fatal call which causes the program to stop.)


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/development@5765 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 4ccc6566df4ff46a6d079b5f8c1fc31e741f039f