Updated the mpas_ocn to be consistent with the latest mpas_atm. The mpas_atm has seen a lot of work, the mpas_ocn should be considered a stub at this time.
There are a LOT of references to atmospheric items, I have
commented most of these out and put in fatal errors should these
be encountered.

I reduced the model_interpolate() routine to a stub just to get
things going. Not sure how applicable the atmospheric routine
is to the ocean. The mpas_atm/model_mod:model_interpolate() is
untouched. We can borrow from there as need be.

The code compiles with ifort (IFORT) 11.1 2010080 and throws
a ton of unused variables etc. It is anticipated there will be
a lot of shared code between the mpas_atm and mpas_ocn. When
we consolidate the common components - we can cut out the dead wood.
 


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/development@5899 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: fa67f65f3b651ae0c98858309765597b61a63f3d