match changes in trunk/kodiak:
removing an old binary restart file and unused subdirectory.
this version of the bgrid code does not support the reading 
and writing of fms-type restart files without the MPI implementation.  
the grid sizes, levels, etc are specified in a namelist 
and the initial data values are taken from the dart IC files.

git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/development@5939 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: b0a11a7f3f558f9e6c2ef768b4f1aba9db87d0a6