These successfully advanced and assimilated with a 4-member ensemble. Tonight the test is with a 30-member ensemble. There will likely be some tuning for the task geometry in the assimilate.csh script that is currently not present.
The Precip_template file is currently not used because the CAM DATM files
have the precip and nonSolar variables on the same 6hr schedule and are
included in the same stream file. Apparently there is/may be some cost
with opening a (redundant) set of stream files. The DART CESM setup
only specifies the Solar_template (daily) and nonSolarNonPrecip (6hrly)
stream text files and does not copy the Precip_template into place.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/development@5960 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: de23d1ec25b256c86a2e30943ef317f9d781b637