This configuration runs a CESM1_1_1 F_AMIP_CAM5 compset on yellowstone using some of Kevin's ccsm 4 restart files for October 2008 generated on bluefire. There are a lot of customizations to make it work for 2008 since the default datasets extend through 2005.
The configuration has not been tested for cycling. Need to ensure the RUN_STARTDATE
and START_TOD in the env_run.xml are sufficient - and that the CONTINUE_RUN = FALSE
is enough.

The short-term archiver has not been tested.

Not sure if there are not efficiencies to be had by using some SourceMods to inhibit
CAM (and other model components) from writing unneccesary files.

Not sure if we can turn the River Runoff model (ROF) off (bug listed in CESM1_1_1 for
some/all F compsets).

Not sure how many ICE instances we need - can we get by with 1?

But - since this was such a pain to get to this point - it needed to be committed for safekeeping.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/development@5981 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: f9578ab6be503ecc789592b28845143ab2d0aefe