These scripts subtract the number of observations with DART QCs of 5 and 6 from the number of observations possible. QCs of 5,6 are namelist settings that FORCE the observation to not be assimilated, so this is a more meaningful representation of the possible observations.
They also plot only 1 region per figure instead of trying to get clever
with 2 or 3 or 4 or ???

The line types and symbols are now consistent across the routines.

The number of observations used is now represented by an asterisk
instead of a plus symbol to avoid confusion with the plus symbol 
occasionally used as part of some lines. 

There is also a quick example of how to simply plot the locations in read_obs_netcdf.m
and plot_obs_netcdf.m



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/DWL@7332 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: d3fcab35bafdbf2ddaf59b810437a0fa99d1bf3c