There are two separate radiative transfer modules, with the possibility of more. This also incorporates the same variable parsing syntax as from the trunk. The Tb branch was a prototype for this, the trunk underwent a code review and renamed some of the variables and keyword values. Those new variable names and keyword values (for model_nml:clm_variables) are now those that were approved for use on the trunk.
The rtm_memls_ss_mod.f90 and rtm_memls_ssva_mod.f90 modules
have many similarities, but enough differences to warrant distinct modules.
It would be nice to a single set of support routines, but that is beyond
the scope of what I should be doing. They are user-contributed code.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/Tb@7357 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 281636826f899ff02bc5a595f97c320349868597