Adding the template information for the UCOAM ocean component "GCOM". At present, the netCDF restart file format is under development, so I am making some assumptions.
GCOM is a regional model. According to Mariangel Garcia,
* primitive 3D Navier-Stokes using Boussinesq approximation
* general curvilinear coordinates
* fully non-hydrostatic pressure equation
* uses UNESCO Equation of State for density

At this point, I don't really have a good grasp of the coordinate information.
"Equations are transformed from the physical grid to an equidistant cube 
and the calculations are performed in that (equidistant) domain."
Someone at SDSU will have to provide the xyz2latlondepth() function.
The observations in the WOD are in lat/lon ...



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/UCOAM@7396 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 9a1cfd911c4783513b5a01686b6af0c6ce116e14