The model_mod.f90 is reordered so that all the required public routines come first, the optional public routines come next, and the module private routines are after that.
I am trying to use the DART style guide and doxygen-friendly comments.

I have not rewritten any of the GCOM support source files to be consistent
with the style guide.

I have added TJH FIXME routines for all the public routines that need to be tested.
That is - ALL of them.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/UCOAM@7402 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 86a759c06413d8858a6f1078bfaea2e00c472017