The get_close_obs() strategy correctly identifies the closest state location. I have a simple inverse distance squared interpolation routine that correctly interpolates to the desired location. The number of nearest neighbors is a problem. Since the vertical resolution is so much finer than the horizontal, a 'realistic' vert_normalization_height (to convert meters to radians) results in all the vertical levels being closer than the next horizontal gridcell. setting debug > 2 will highlight the effect.
I tinkered with vert_normalization_height till I got a value that generated nearest neighbors in the horizontal for the interpolation.  For testing, I could/should just use
the value at the nearest point for my interpolation result. There is much more to do
and the GCOM people should be doing this anyway.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/UCOAM@7633 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 3e7f33d290e314924589e2edb116ac52b24b85e6