These are the checks and documentation that explain the relationship between 'cutoff' and 'vert_normalization_height'.  Basically, with a warped grid, the vertical exaggeration needs to be tailored to the relationship between the horizontal grid spacing and the vertical grid spacing. The model_mod.html explains this in great detail.
I removed a few FIXME's that I checked, I removed an entire untested routine
that wrote a separate netcdf file for _some_ of the grid variables. The
routine was left over from POP.

The only user-visible change is to the namelist. Given that GCOM
is a high-res model - I put in a default value for vert_normalization_height
that is appropriate for the ~ 50m horizontal grid spacing and a vertical
grid spacing of ~ 10 meters or so. This is the example that generated the figure
with the 'local' innovations.

The plot_grid.m function was modified from its POP ancestry to work with GCOM.
It was used to generate the grid figure. It does more.
 


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/UCOAM@8055 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 814fa973c952c86cdfa46ee6a0df6509853edf3a