Reading the land mask and using it to create the tile fractions for just the land gridcells. The restart file variables are only 'land' cells, but the tile_fractions array is physically-shaped (as are all the 'input grid' variables). I needed a way to relate the input grid variables to the 'land-only' (model) variables. The land mask is specified in the model_grid.nml. Can now work on writing the model_interpolate() code. Finally.
I'd like to propose that we always use

   model_grid.nml:jules_model_grid:land_only     = .true.         and
   model_grid.nml:jules_model_grid:force_1d_grid = .true.

to reduce permutations. The model grid will always be 'sparse', the input
grids (containing the tile_fractions) will always be 2D - so a land mask
will always be needed to resolve the two.  One nuisance - the
restart files have dimension 'land' - the output file has dimension 'x',
the land mask and tile fractions have dimensions Nx,Ny - but Nx is not 'x',
and 'x' is 'land' ... bookkeeping nuisance in the DART diagnostic files.

At the moment, this is not enforced in the setup scripts 
(which seems like the logical place to do it).



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/jules@8472 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: c702a6ea89a3d9f925c3039cef16fd2019dcd032