model_interpolate is now doing vertical interpolation. The aggregation for PFTs/tiles is not complete. The unused interfaces compute_gridcell_value() and get_grid_vertval() have been removed - as have all the unused variables and a couple other stub functions.
The loc_get_close_obs() routine that checks to see which sparse
locations are close to an arbitrary location should be checked out
a bit more. So far, the test has been easy, the target location is
almost right on top of a sparse gridcell. Need to test if the location
is farther away, on an edge, too far away, etc. - SHAMS?



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/jules@8569 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 2d2e543f332ef28bd34d3fc7b28f474238bbbc42