Renamed land_fractions to land_mask to match the JULES nomenclature. The units are actually fractions, so ...
I've confirmed that if you stride through all the longitudes first (fastest), then
the latitudes, then ... that you match the unraveling done in JULES when they
strip out the gridcells that do not contain land.

The DART netCDF diagnostic files are believed to be done and correct - caveat
the fact that we may want to add more variables that relate 'land' index to some
geographical i,j.

I put in a check to ERROR OUT if the 'y' dimension in the JULES "dump" or "hour" 
files is not 1. Making the command decision that Nland MUST EQUAL Nx.

Still have to create the lookup table arrays for the land to i,j and vice-versa.
Still have to implement the better/faster model_interpolate that can exploit the
physical grid information to determine if the closest gridcell is not land.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/jules@9362 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: ede838fe62e2e982448a0de904f57bb5a626cb2a