Import the style guide changes, now that a base line model_mod.f90, which came out of the process of merging branches/cam onto the trunk, has been committed to the trunk.  A large fraction of the changes are to white space, so turn all white space differences off in xxdiff. Most of the rest are minor variable name changes, e.g. going from CamelBack to under_score, and the inclusion of a new comment character, '!>', to enable doxygen to harvest documentation comments from the fortran code.
There are updates to the innovation damping algorithm for the layers
near the model top, denoted by 'Better damping'.  This is an improvement
for vert_coord = 'pressure', but is absolutely necessary for
vert_coord = 'log_invP'.  On the other hand, using the default CAM damping
div24del2flag = 4 removes the need to damp innovations near the top
or limit the height of observations used.

There are a couple important bug fixes in get_val_height, model_heights, dcz2,
and set_ps_arrays, which affect only assimilations of obs on heights (e.g. GPS).




git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/trunk_with_cam_style@9476 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: d6d37f839ebb80885d07ea5943215155d07374b6