The model_interpolate routine is tested and believed to be working. The only possible extension to it is to check if the scpool dimension is 1 and then grab the value ... but if the scpool dimension is 2, I have no idea what to do. so - until I get more instruction about scpool - it simply throws a fatal error (rather than silently fail as model_interpolate is allowed to do).
The model_mod_check program now actually honors the kind_of_interest specified
in the namelist.

The set_grid_boundary() routine still needs to be checked for wrapping 
around the prime meridian - particularly when it tries to calculate the 
grid width so it can divide it by two and add/subtract to the highest/lowest
longitude center.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/jules@9526 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: bfb0b58e7cecb65c6ef641a533641ca30255a34e