Branch to fix CLM restarts that are not bitwise with the trunk.
The reason this happens is because the CLM model_mod modifies the 
state with missing_r8 values where information is not used in 
clm_to_dart_state_vector.  This routine does not get called when 
reading directly from netcdf so state values are not updated.  

Similarly when using direct_netcdf_write missing values are not
updated with original restart values ( as in the trunk ).

The aim of this branch is to find an elegant solution to have
restart files that are bitwise with the trunk!



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_fix_clm_restarts@9634 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: a9f069ff69fd722753a0125a2b359b0bb03058f4