update alicia's code:  some compilers don't like  if (logical == .false.) then it needs to be either:  if (.not. logical) then or  if (logical .neqv. .false) then i prefer the first one.
also use our file open and close routines which
assign a free unit number instead of hardcoding
one in the code.  indent logic.  move comments
up and make them in doxygen format.

add the new module to the path_names files,
and make a first stab at calling the code in
the get_close_obs() and get_close_state() code
in the CESM model_mod.

untested as of now.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/xcesm@9725 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 46ef5021afd1ec60fcf4c64c353e2d4f9cc61d60