General code cleanup. Using error_handler for most of the messages now. Pushed the find_closest_surface_location() code out of model_interpolate() and into find_closest_surface_location()
Removed the nuisance code from get_close_obs(). Technically, this is
exactly the same as the one in the location module, but I left the code
in for future use if anyone wanted to perform special localization (as in POP).

The feom_modules module was using hardcoded units for reading, etc. and was 
stealing unit 10 from the logfileunit. replaced hardcoded units with a call
to DART's get_unit() function - it was already calling do_output() from
our utilities_mod. 

Fixed the nc_write_model_[atts/vars] routines so we can write out
the prognostic variables - kinda - no info as to how to parse the 3D vars
back into 2D,level vars.

fixed an error in model_mod_check ... was not allocating state_vector
after the model_size was calculated. Runs up through test 7 ... have not tried
anything more exhaustive.

For some reason, when using the null_mpi_utilities_mod.f90 ... the call to
my_task_id() is hanging the model_mod.f90 ... 

Removed all mention of MPAS (I think).

Added the SSH to the state so we can test the parts that query 2D vars as well as 3D vars.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/FEOM@10100 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 8c5a00cbd940fcdc093ccb0d78a1ae75e49d5639