Using the coord_nod3D array from FEOM for all get_state_meta_data purposes. The dump_tables() routine confirms that get_state_meta_data is returning locations that identically match those in coord_nod3D.
dump_tables() is only turned on by hardcoding 'state_table_needed = .true.'
since it is exceptionally expensive to write out a 'rosetta-stone-like' file
that has *every* DART state vector index and locations as well as those
for the various subsets, etc. needed to compare things.

With a cutoff of 0.00005 ... the obs_seq.final had identical priors and 
posteriors for all three observations, but there changes to the state - 
in the wrong place. Since I believe get_state_meta_data() is correct - 
this might mean the model_interpolate is getting the wrong part of the state
for the forward operator.
 


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/FEOM@10244 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: cb3e831cf1e1e1d16414cc0347b51518a6be81ba