common bitwise besting routines added to directory bitwise.
start of shell script to do bitwise test for wrf. currently
the script runs a quick build on rma_trunk and trunk for wrf.
sets up a run directory to do the bitwise test and links the
appropriate files in the run directory.  The test then converts
the test_trunk restarts to model files for testing.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_bitwise@10289 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: d6527cf0358918e449a64f3eb83e38e5c3eaa5ff