The variables in the initial conditions files and the DART diagnostic files have been made consistent. The use of these files is untested, but will be after we come back from Du's farewell lunch.
The namelists and existing diagnostic files have been changed to reflect 
the new variable names. The DART diagnostic scripts must also be changed
to accomodate these new variable names.

The scripts that contain the NCO commands to manipulate the variable
names, dimensions, attributes, etc. are committed, but the usage notes may
not be correct just yet.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/pda@10586 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 30acd5f22db6467e132fe970a05046d9c540b62d