This configuration was bitwise identical. Implemented trunk revisions to the wrf model_mod.f90 to the wrf_chem/model_mod.f90 The trunk revisions were r9415 and r10379 - both pertaining to the incorrect indexing if domains had different variables. This problem was documented in Jira issue DARTSUP-295
This uses the smaller observation sequence file so the filter can run in ~ 8 minutes.
Because it can be run faster, the wall-clock time in the run script section was reduced
in hopes of getting to run sooner. The script also sends mail to the user when the job
is completed.

The run script is designed to exit after filter and before advancing wrf_chem.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/mizzi@10925 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 0ae1a2617b631946a391275cc5266b9ec40e2dc9