This test is bitwise identical.
This implements trunk  wrf/model_mod.f90 changes in r8397 
to the wrf_chem/model_mod.f90

Paraphrasing from r8397:
"This fixes JIRA ticket DARTSUP-247 and addresses DARTSUP-245
We believe it fixes the original issue in DARTSUP-245, but are
awaiting confirmation that it fixes the subsequent issues in the ticket.
DARTSUP-245 pertains to the ability to interpolate the land mask.
DARTSUP-247 is the shape of the 2D variables in the DART diagnostic files."




git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/mizzi@10926 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 710e78941582c67c0eb723c75673ac24bb8a1d9b