Merging in changes from the trunk that have tested OK or are changes to unrelated code. There are still some bitwise differences in the output prior inflation file that must be coming from the mpi_utilities_mod.f90, adaptive_inflate_mod.f90, assim_model_mod.f90, or assim_tools_mod.f90.
The output files reproduce exactly, as do the diagnostic files.
There are two changes to some obs_def modules that cause the observation sequence 
header being tested to have different integer indices, but the data values are identical.

As soon as I identify the source of the bitwise differences, I will commit the rest of the
files from the trunk. There are some changes to the MADIS converters that I 
want to discuss before committing them.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/mizzi@10931 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 64818908d1b3921346e039f99598312962e3b541