make several routines take an ensemble size and compute multiple values in one call.
add staggering to the quad routines so you can get
2d fields underneath U and V locations.

start to streamline the height column routines and
supporting subroutines.

change the interface to the quad code for regular
grids.  return an array of 4 lon/lat indices, so
subsequent code can do i=1,4 and pick up the right
pairs of values in the right order, so when you call
the evaluate routine you have the field values in
the right order.

also add some doxygen and fix up some misleading
comments.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/recam@12172 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 8e0ad36699aa7ee5689278bc4e7a1b816c1ab372