the NAG compiler pointed out that if you don't call initialize_utilities() before calling register_module() you get a recursive call to the init routine.  i changed the code so anything used BY the init routine cannot be called first - thus avoiding recursion.  i also removed some unused code, changed some of the error handling so it's clearer when you're writing an error message directly to the stdout and when you're writing to a log file.
the netcdf_utilities_mod has changes that try to cache
the last N filenames, and match up the ncid values to the
name so the caller doesn't have to keep passing in the
filename just for error conditions.  if this works, this
should be migrated to the rma_trunk as well.

renamed a bunch of the interfaces to try to be more
helpful to a non-netcdf user.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/recam@12224 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: a6868ddd3d611a202f9fd3b3a8a42b7c873bf4d7