update fixsystem to sed in and out the lines that differ for the NAG compiler.  it's available on the CGD machines and it has found several bugs/questionable code.  we should be testing here more frequently, thus i'm making it easier to just check out and build there.  all the system differences that are incompatible with other compilers are currently isolated here.
one other difference, just for the record, is that nag doesn't
use 8 as the kind for real(r8) - so code that assumes 4 and 8
are real kinds doesn't work here.  we don't have that in any
of our modules, but wrf had some wrf source code that made
this assumption.  i'll be committing that in a bit.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/recam@12273 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 246aeff7df815f21f42b13571249b788cb79174d