this has notes and partially implmented code for putting 2 different additional distribution types into an ensemble type. also an option on transpose_type 3 for replicating across a subset of the total tasks.
this requires a group to be created in the mpi_utilities
and for the window manager to know which communicator to use.
can we make a group communicator that defaults to using all 
N tasks so the window manager can always just use the group
communicator for the ensemble mean?



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_distrib@12425 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 4db0d5a82d73185d144cc0b98d0becbdc8b5430d