CLM5_setup_spinup is working - but it is using netcdf, not pnetcdf since pnetcdf crashes because the finidat interpolation filename is not multi-instance friendly. Don't know why, but netcdf seems to not crash - it is a reported bug.
CESM2_0_setup_hybrid starts from restart files that do not need to be interpolated,
so it can use pnetcdf. I am using it to test if you can run-time switch between
pnetcdf and netcdf. If possible, I'd like to change the continuation runs of the
spinup to use pnetcdf since those runs do not need to be interpolated.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/cesm_clm@12455 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 1964c88b7f84c3041e5a3165185acb73a5743734