W Douglas Cramer 3:45 PM (21 hours ago)
I've completed the scripts that create and execute the ensemble openggcm runs, as well as a test script that simulates the interaction with filter. All of the scripts are in the rma_openggcm/models/openggcm/shell_scripts directory of the dart account.

create_base.csh : This creates the base openggcm run, which includes building the libraries and saving source files for the ensemble runs.

create_runs.csh : This creates the ensemble runs and the build_all and submit_all scripts. Solar wind input data and model parameters can be changed between the runs by editing the swdata and runme files, respectively.

dart_test.csh : This simulates the interaction with dart. It is executed with the ensemble runs by the submit_all script.

More work could be done on these, but they perform their function for now.

Doug



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/openggcm@12482 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 99f4e47041eb9f4f897029c534be3016b0fff288