The observation converter seems to be working. Turns out the two sensors are 8cm apart. Prabhakar decided if they both are 'good', we use the average - if only one is 'good', we use only it.
The definition of 'good' comes from the qcdim of the qc flag.
The interpretation of the integers come from a dump of the
processing_status and quality_flag variables.   At present, I am using
a hardcoded integer ... should really search for the integer pertaining
to the desired string ...

Needs documentation.


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/lanai_terrsysmp@12496 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: e59a0a434ca3c5486d06ffca58aad24f229a9e26