The observations are actually soil water content, which is a percentage. To get soil moisture, you have to multiply by porosity.
Worse ... the Intel compiler and gfortran interpret the NaN value _very_ differently.
The Intel compiler (with debug options) simply core dumps uselessly when it reads the NaN.
The Intel compiler (with moderate optimization) blows past this and converts TWICE as
many observations as the gfortran compiler.

gfortran with debug on or off seems to not blow up ... who knows what is right ...




git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/lanai_terrsysmp@12509 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 70bff757fe54735ec75873d4e8d6ff6358a34d53