updates to the recam branch:
- merge in recent changes from the rma_trunk, including
the cleanup of KIND_ to QTY_ in many random files.

- remove lanai bitwise option for perturbing a single
initial state into an ensemble.  this was removed from
POP, WRF, MPAS_ATM, and bgrid.  also assim_tools_mod.

- use netcdf utility routines in the POP nc_write_model_atts()

- fix the perturb routines in wrf and mpas_atm.  they are
rarely used but had indexing problems with the variable counts.

- more routines added to the netcdf utility module

- added get_maxdist() routine to all location modules.
model_mods can now query for the cutoff distances.  note
that this will be 2*cutoff - the full localization distance -
because that's what's stored in the modules.  the routine
takes a gc (get close type) and an obs type as arguments
because one of the modules (threed_sphere) allows different
cutoffs by obs type.

- update the cam model_mod to include a new namelist item:
         no_normalization_of_scale_heights
if true, it does NOT normalize the scale height by the
surface pressure.  if false, it is backwards compatible
with previous implementations.  i have this code for
wrf and mpas_atm but it's not part of this commit yet.

- adaptive inflate has code to try to turn the combination
of namelist options into an english language description of
what options you've selected.  due to my misunderstanding
of how this works i was doing the wrong test for when the
variance was adapting.  this is a log message change only;
no results change.


i believe all these changes should eventually end up on
the rma_trunk.  once the recam branch has been tested and
seems to be working ok, these can all merge onto the rma_trunk
and eventually into manhattan.

still left to do:  see if we can set the top-of-model-ramp
by model levels, and disable it when horizontal distance
only is selected.




git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/recam@12563 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 4aecd43660c4ed9ae69e7bbf94ace066b5789bc1