Support DART QC = 8
The html documentation is more correct.
The html files are structured very similarly to enable spotting differences.
The source codes are structured very similarly to enable spotting differences.

Should reorder the routines in the 3D codes to be ordered like in the 1D code
to facilitate differencing, but have not done so yet.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/gigg@12620 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 8d68c726e82ae89e734cb453b1086a0de59f1d65