fix a bug in the 'no normalized scale height' for surface obs.
change how you specify the top ramp to use model
levels instead of pressure, and only specify the
top of the ramp.  the bottom will be computed
based on the cutoff distance.  this will result in
different specific observation increments but the
hope is that overall the results are the same or better.

change the names (only) of the global_xxx variables
to be ref_xxx because they're the reference values.

i've only done simple tests on these - i'm committing
the code now so i can move to cheyenne for larger and
more complicated tests.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/recam@12698 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 0663b24caf2c1a3c5d178ace44612a166b57b250