The assimilate.csh.template file has been modified to properly archive DART diagnostic files under ?all? domain configurations.
Reference the new filesystem layout on Cheyenne.
The freerun and spinup scripts should be consolidated - the terminology is confusing.
We generally start from a single CLM state that is already 'spun up'; we are taking
that single state and performing a 'free run' where that state is forced by an
ensemble of data atmospheres.

Some of the setup scripts have been modified to work with DATM files at a different
resolution than the CLM resolution. These are generally the freerun versions.
It is presumed that the assimilation experiments will stage output from a freerun,
so a change of resolution is not natively supported.

The CLM5_setup_pmo is being modified to use a separate parameter file that will have
all the user-specified variables in one place. Under Development.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/cesm_clm@12900 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: f21a476008c7f90a54ca29beb9dece4868f5ff27