This checks each gridcell for outliers and replaces them with a noisy estimate.
The methodology is that all the ensemble members for that gridcell and at a particular time
have their variance calculated. The largest value is replaced by the median and the variance
is recalculated. If the ratio of the variances exceeds some threshhold (in this case a factor of 100.0)
the largest value is considered to be an outlier and is replaced by the value at the 75th quantile 
that has a random_gaussian() noise from a distribution with the interquartile range as the stddev.
The process is iterated until there are no outliers in the gridcell. Preliminary tests for all of 
2008 with the downwelling longwave radiation indicates that _at_most_ there were 3 outliers 
in a gridcell. This is loosely based on the Modified Thompson Tau test.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/cesm_clm@12901 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: f3d7ceb674ddf7324825145e48573925787729c0