change name of variable 'cutoff' to 'vert_localize_dist' to make it more clear that this distance has already been multiplied by 2. it isn't the cutoff anymore; the name was misleading.  note this is just a variable name change - nothing in the computation has changed.
also changed the subroutine name v_minus() to v_down() to reflect
that this function returns the new vertical location if you start
at vertical location 'a' and move 'b' units towards the earth surface.
calling it 'minus' was confusing because in all cases except height
you are adding b to a to move lower in the atmosphere.

tldr: only name changes for code clarity; no algorithmic changes.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/recam@12908 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 584245529308f28e4f8d2f006c32f70e899d5f20