The progvar structure is completely gone. The natural _FillValue and missing_value values are intrinsically restored in the direct_netcdf_mod.f90 since those values are naturally part of the state structure. Technically we should check the variables in the output files to see if they are compatible with the state_structure info.
Removed the need to call 'dart_to_model' ... it still may be needed when we
do snow DA to repartition SWE.  With CLM5 the BandDiagonal solver is choking
when we attempt to update soil moisture. This is a problem we must solve.

Some (but not all) of the setup scripts use the DART_params.csh parameter file.
There are still some values in it that may not belong ... feedback from people
will be useful to determine what should go where.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/cesm_clm@12927 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: b12734537510569fdf30c8552f0b9ec669ff17f9