Removed the need to explicitly specify an initial interpolation filename as it is not needed in the release versions.
Uses the DART_params.csh file and overwrites some of the variables
explicitly in the script as it makes sense. Those variables are not
used by assimilate.csh or perfect_model.csh (They source DART_params.csh,
so there could have been a conflict.)

Removing/Deleting CESM2_0_setup_spinup as 'spinup' is a misleading term.
Technically, we are taking an already spun-up state and starting a 
free run.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/cesm_clm@12928 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 77d073b967feba3bc0871b5d07289a03df9f851b