remove the allocated space for obs_seq.final from all tasks other than the one which is going to collect the data and write that file. for large ensemble sizes where all the obs values are being output, this is a big memory savings.
also add the option to avoid computing the posterior
forward operator values, and don't allocate space in
the obs_seq.final for posteriors.  this should be
both a time and space savings.

add a check so users can't ask for both posterior
inflation and to not compute posteriors.

add some code in filter as a preparation for streamlining
the trace and timing messages to try to make the code
easier to read.

update the dopplerfold version of filter to stay in sync
with the changes to filter, even though this is only
used (as far as i know) with the wrf model.  still,
this is a proposed version to be migrated to the
rma_trunk and on to Manhattan, so update it now.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/recam@12964 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: d678f5e55af768755027860cca8a7c94c5fa75b4