Can increment/decrement dates by months ... dates near the end of the months are problematic. Adding a month to the last day of a month will result in the last day of the next month. Subtracting a month from the last day of a month will result in the last day of the previous month.
This logic renders some dates unreachable or overloaded. For a more thorough set of examples, turn on 'verbose', set the bin_separation months to be /= 0 and a table
will be generated in the output (date_increment_test() will be run).

Jun 30 + 1 month results in July 31, not July 30
Mar 31 - 1 month      results in Feb 29 or 28
Mar 30 - 1 month also results in Feb 28 or 29
Mar 29 - 1 month also results in Feb 28 or 29
Mar 28 - 1 month also results in Feb 28



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/cesm_clm@12990 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 22b8f6bf22f56d3b9d2772cad391684e2840aa9d