Added missing vertical coordinate information for (?,<anylevelcoordinate>,lat,lon) variables. There are multiple vertical coordinate possibilities, levgrnd, levsoi, levtot, ... Added code to use any variable explicitly stated as QTY_SOIL_MOISTURE as opposed to adding QTY_SOIL_LIQUID_WATER and QTY_SOIL_ICE together.
The gridded history variable H2OSOI is QTY_SOIL_MOISTURE and has units mm3/mm3
The restart variables H2OSOI_LIQ and H2OSOI_ICE  have units kg/m2

The satellite soil moistures typically come as m3/m3 so there would seem to
be no units conversion necessary if using H2OSOI for QTY_SOIL_MOISTURE.
Some satellite soil moistures come as percentages, these have to be divided by 100.0
to have the same units as H2OSOI. 

One of the benefits here is that the COSMOS operator is written in m3/m3 ...
It was always a pain to convert kg/m2 to m3/m3

The downside is that H2OSOI is only available at the gridcell level.
If more accurate forward operators are desired (for the column level)
then more work is needed.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/cesm_clm@13185 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 9ed5572c4b32f6c17a92c6ee542d94a1d9e56c10