add an option to allow a 1D latitude array to start at 90 and to go -90 instead of starting at -90 and going to 90.  this entails iterating the 1d array in the opposite order, but leaves the rest of the code alone because the return values are the corner indices of the right quad, and the fraction across the quad.  as long as we're returning the right corner indicies, the interpolation code doesn't care which order the latitudes are in.
also cleaned up some comments and start to explore
exactly what would happen with a grid that is specified
as -180 to 180 in longitudes.  i believe adding 360 to
all the negative values makes this valid, since we
already have code that knows how to interpolate at
the "seam" between a line of grid cells that ends
around 360 on the left and starts around 0 on the right.
(add 360 to the "0" side and compute the fraction across
like normal).



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/basic_earth@13206 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 56566a2d29711e450856b5fd33836d5a5fe72a1c