Successful assimilation of navdas obs ... kinda
The navdas obs were from 2013, the state is for 2018 - I shifted
the obs times to match the model time and subsetted to [200 260 20 60]
Still some vertical rejection, the expected outlier rejection.

The COAMPS state lives as sigma levels in the DART world with locations
that are the actual sigma values as opposed to level indices.

The fill_inflation_restart tool is having trouble with the mean exner
functions ... they are marked 'noupdate' by coamps_statevar_mod.f90
which prevents the inflation tool from creating them - however, since
they are part of the DART state - they are required.