FPAR forward operator, new CAM DATM files, obs_loop
The CAM DATM files execute, but the time offset needs to be explored.
The new DATM files are at 1hr and 3hr, the template files were based on
the old DATM files at 6hr and 1day.

The obs_loop program is used to modify existing observation sequence files to
a new date. I wonder if create_fixed_network sequence might not do exactly
the same thing ...